lambdas/interview/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
boto3==1.34.14
botocore==1.34.14
certifi==2023.11.17
charset-normalizer==3.3.2
decorator==4.4.2
distro==1.9.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
jmespath==1.0.1
moviepy==1.0.3
numpy==1.26.3
openai==1.6.1
pillow==10.2.0
proglog==0.1.10
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==1.26.18
